requirements.txt
Django==2.2.5
celery==4.4.1
redis==3.4.1
djangorestframework==3.10.2
requests==2.22.0
djangorestframework-api-key==2.0.0
psycopg2>=2.7.5,<2.8.0